requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

torch>=2.2.0
transformers>=4.41.0
accelerate>=0.30.0

soundfile==0.12.1
librosa==0.10.2.post1
numpy>=1.26.0

jiwer==3.0.4

pyyaml==6.0.2
pydantic==2.10.3
tqdm==4.66.5
pandas==2.2.3